requirements.txt
python-telegram-bot==20.0
flask
requests
python-dotenv  # For handling environment variables securely
